Move TestDataPreparer to Flink and Lite engines
